+++ server/src/main/java/backend/server/service/user/UserServiceImpl.java
@@ -8,6 +8,10 @@ import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -19,6 +23,7 @@ import java.util.UUID;
 public class UserServiceImpl implements UserService {
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
+    private final AuthenticationManager authenticationManager;
 
     @PersistenceContext
     private final EntityManager entityManager;
@@ -49,18 +54,63 @@ public class UserServiceImpl implements UserService {
         UUID id = userRequest.getId();
         User user = findById(id);
 
+        boolean authorized = user.getUsername().equals(SecurityContextHolder.getContext().getAuthentication().getName());
+        if (!authorized) {
+            throw new RestException(
+                    HttpStatus.UNAUTHORIZED,
+                    "Unauthorized to update user with id " + id,
+                    System.currentTimeMillis()
+            );
+        }
+
+        if (userRequest.isPasswordIsChanged()) {
+            if (userRequest.getOldPassword() == null || userRequest.getOldPassword().isEmpty() || userRequest.getOldPassword().length() < 6) {
+                throw new RestException(
+                        HttpStatus.BAD_REQUEST,
+                        "Current password is invalid",
+                        System.currentTimeMillis()
+                );
+            }
+
+            authenticationManager.authenticate(
+                    new UsernamePasswordAuthenticationToken(
+                            userRequest.getUsername(),
+                            userRequest.getOldPassword()
+                    )
+            );
+
+            if (userRequest.getNewPassword() == null || userRequest.getNewPassword().isEmpty() || userRequest.getNewPassword().length() < 6) {
+                throw new RestException(
+                        HttpStatus.BAD_REQUEST,
+                        "New password is invalid",
+                        System.currentTimeMillis()
+                );
+            }
+            user.setPassword(passwordEncoder.encode(userRequest.getNewPassword()));
+        }
+
         user.setFirstName(userRequest.getFirstName());
         user.setLastName(userRequest.getLastName());
         user.setUsername(userRequest.getUsername());
-        user.setPassword(passwordEncoder.encode(userRequest.getPassword()));
         user.setRole(userRequest.getRole());
         user.setEmail(userRequest.getEmail());
+
         return userRepository.save(user);
     }
 
     @Override
     public void deleteById(UUID id) {
-        findById(id);
+        User user = findById(id);
+
+        boolean authorized = user.getUsername().equals(SecurityContextHolder.getContext().getAuthentication().getName())
+                || SecurityContextHolder.getContext().getAuthentication().getAuthorities().contains(new SimpleGrantedAuthority("ADMIN"));
+        if (!authorized) {
+            throw new RestException(
+                    HttpStatus.UNAUTHORIZED,
+                    "Unauthorized to delete user with id " + id,
+                    System.currentTimeMillis()
+            );
+        }
 
         userRepository.deleteById(id);
     }
@@ -87,7 +137,7 @@ public class UserServiceImpl implements UserService {
     public List<User> findUsersLikedByBlogId(UUID blogId) {
         String query = "SELECT DISTINCT l.user FROM Like l WHERE l.blog = :blogId";
         return entityManager.createQuery(query, User.class)
-                .setParameter("blogId", blogId)
-                .getResultList();
+                            .setParameter("blogId", blogId)
+                            .getResultList();
     }
 }
